fastapi-aph/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
asyncpg==0.29.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
psutil==5.9.6